% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visx_cor.R
\name{as.data.frame.visx_cor}
\alias{as.data.frame.visx_cor}
\title{Convert visx_cor to data.frame}
\usage{
\method{as.data.frame}{visx_cor}(x, ...)
}
\arguments{
\item{x}{a visx_cor object}

\item{...}{additional arguments (ignored)}
}
\value{
a data.frame of pairwise association values
}
\description{
Returns the association value matrix as a data.frame.
}
