% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visx_cor.R
\name{plot.visx_cor}
\alias{plot.visx_cor}
\title{Plot method for visx_cor objects}
\usage{
\method{plot}{visx_cor}(x, min_cor = 0.3, sig.level = 0.05, show_signif = FALSE, ...)
}
\arguments{
\item{x}{a visx_cor object}

\item{min_cor}{minimum association to visualize (default 0.3)}

\item{sig.level}{significance threshold (default 0.05)}

\item{show_signif}{whether to show significance overlay (default FALSE)}

\item{...}{additional arguments passed to \code{\link{npc_mixed_cor}}}
}
\value{
a ggplot object, invisibly
}
\description{
Generates a network plot of correlations and associations.
}
