% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{print.visx_check}
\alias{print.visx_check}
\title{Print method for visx_check objects}
\usage{
\method{print}{visx_check}(x, ...)
}
\arguments{
\item{x}{a visx_check object}

\item{...}{additional arguments (ignored)}
}
\value{
x, invisibly
}
\description{
Print method for visx_check objects
}
