% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visx_cor.R
\name{print.visx_cor}
\alias{print.visx_cor}
\title{Print method for visx_cor objects}
\usage{
\method{print}{visx_cor}(x, ...)
}
\arguments{
\item{x}{a visx_cor object}

\item{...}{additional arguments (ignored)}
}
\value{
x, invisibly
}
\description{
Print method for visx_cor objects
}
