% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coradar.R
\name{print.visx_coradar}
\alias{print.visx_coradar}
\title{Print method for visx_coradar objects}
\usage{
\method{print}{visx_coradar}(x, ...)
}
\arguments{
\item{x}{a visx_coradar object}

\item{...}{additional arguments (ignored)}
}
\value{
x, invisibly
}
\description{
Print method for visx_coradar objects
}
