% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{visx_collapse_levels}
\alias{visx_collapse_levels}
\title{Collapse levels of a categorical variable}
\usage{
visx_collapse_levels(
  df,
  var,
  levels,
  new_label,
  new_type = c("factor", "ordinal")
)
}
\arguments{
\item{df}{a data.frame}

\item{var}{name of the column to collapse}

\item{levels}{character vector of levels to combine}

\item{new_label}{the label for the collapsed levels}

\item{new_type}{type of the resulting variable: "factor" or "ordinal"}
}
\value{
a single-column data.frame with the collapsed variable,
  named \code{<var>_bin}
}
\description{
Collapse levels of a categorical variable
}
\examples{
df <- data.frame(color = c("red", "blue", "green", "red", "blue"))
visx_collapse_levels(df, "color", c("red", "blue"), "warm")

}
