% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{visx_ratio}
\alias{visx_ratio}
\title{Compute ratio of two columns}
\usage{
visx_ratio(df, var1, var2, name = NULL)
}
\arguments{
\item{df}{a data.frame}

\item{var1}{name of numerator column}

\item{var2}{name of denominator column}

\item{name}{name for the new column (defaults to \code{var1_over_var2})}
}
\value{
a single-column data.frame with the ratio
}
\description{
Compute ratio of two columns
}
\examples{
df <- data.frame(a = c(10, 20), b = c(2, 5))
visx_ratio(df, "a", "b")

}
