% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_VisXplore.R
\name{VisXplore}
\alias{VisXplore}
\alias{VisX}
\title{Run the VisXplore application}
\usage{
VisXplore(data)

VisX(data)
}
\arguments{
\item{data}{a data.frame to explore}
}
\value{
Histograms for continuous variables;

Barplots for discrete variables;

Correlation diagram;

Variance inflation factors (VIFs) and R-squared;

Correlation coefficient and p values from correlation test
}
\description{
This function starts the VisXplore application for basic data
pre-processing and variable selection.

\code{VisX()} is a deprecated alias for \code{VisXplore()}.
}
\details{
This application currently supports files in csv format.

Application interface will pop up after calling this function;
data can be uploaded in that interface;
close the interface or interrupt R to stop the application.
}
